+++ addon/globalPlugins/subtitle_reader/primeVideo.py
@@ -8,18 +8,20 @@ class PrimeVideo(SubtitleAlg):
 	def getVideoPlayer(self):
 		obj = self.main.focusObject
 		appName = obj.appModule.appName
-		videoPlayer = find(obj, 'parent', 'class', 'atvwebplayersdk-overlays-container fpqiyer')
-		if videoPlayer:
-			videoPlayer = obj.parent.parent.parent.parent
+		videoPlayer = find(obj, 'parent', 'role', role('document'))
+		videoPlayer = videoPlayer.firstChild
+		if videoPlayer.IA2Attributes.get('class') != 'webPlayerSDKContainer':
+			videoPlayer = videoPlayer.lastChild.firstChild
 		
-		return videoPlayer
+		return videoPlayer if videoPlayer.IA2Attributes.get('class') == 'webPlayerSDKContainer' else None
 	
 	def getSubtitleContainer(self):
 		videoPlayer = self.main.videoPlayer
-		container = videoPlayer.next.firstChild.firstChild
+		container = videoPlayer.lastChild.firstChild.firstChild.lastChild.previous.previous.firstChild.firstChild.firstChild
 		return container
 	
 	def getSubtitle(self):
 		obj = self.main.subtitleContainer
+		obj = obj.next
 		return super(PrimeVideo, self).getSubtitle(obj)